requirements.txt
tensorflow
pandas
numpy
matplotlib
pyYAML
Flask
-e .